Don't check Content-Length header anymore
Because it is not present in Flask 0.11. We keep merging the headers from the `headers` variable with the headers from the `response.headers`, because changing the response may change the headers, especially the `Content-Length` header. In Flask 0.11 this `Content-Length` header is no longer present in the `response.headers`, so we check whether the `response.data` has changed.

Fixes #290